requirements.txt
asgiref==3.8.1
attrs==24.2.0
autopep8==1.6.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==7.0.0
Django==4.2
django-allauth==65.0.2
django-cors-headers==4.4.0
django-environ==0.8.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
gunicorn==20.1.0
idna==3.8
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mysqlclient==2.2.4
oauthlib==3.2.2
pillow==10.4.0
pycodestyle==2.8.0
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0.2
channels==4.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.1
setuptools==75.2.0
sqlparse==0.5.1
toml==0.10.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==5.3.0
djangorestframework==3.15.2
drf-spectacular
uvicorn[standard]==0.32.1
gunicorn==20.1.0
numpy==1.24.4
pandas==2.0.3
pandas_ta==0.3.14b0
statsmodels==0.14.1
channels_redis==4.2.1